Move config parsing
Moves the config parsing method, so that other dependents have an easier time using it
